debug

+++ cambrian/model/cambrian_arch.py
@@ -213,7 +213,7 @@ class CambrianMetaModel:
 		if pretrain_mm_mlp_adapter is not None:
 			mm_projector_weights = torch.load(pretrain_mm_mlp_adapter, map_location='cpu')
 			def get_w(weights, keyword):
-				return {k.split(keyword + '.')[1]: v for k, v in weights.items() if keyword+'.' in k}
+				return {k.split(keyword + '.')[1]: v.to(torch.float) for k, v in weights.items() if keyword+'.' in k}
 
 			self.mm_projector.load_state_dict(get_w(mm_projector_weights, 'mm_projector'),strict=True)
